mulistgple {mulistgple}	R Documentation

Evaluation of multiple list

Description

Evaluation of population size based on multiple list experiments.

Usage



mulistgple(DATA)


Arguments

DATA
a list whose elements including: list with covariates of each primary period, list with multilist observe of each primary period, vector with observed numbers of each primary period, initial vector value

Details

nothing!

Value

return a list, whose elentments including: the estimates of parameters, the estimates of population size, the standard deviation of parameters, the standard deviation of population size.

Note

nothing

Author(s)

Liu Wei

Examples


rm(list=ls(all=TRUE))
library(MASS)
v=1000
T=5
d=4
m<-5
set.seed(1234)
DATA <-gendatafun(v,T,d,m)
estparsd<-mulistgple(DATA)

[Package mulistgple version 0.1.0 Index]